+++ index.html
@@ -6,7 +6,6 @@
   <link rel="stylesheet" href="css/normalize.css">
   <link rel="stylesheet" href="css/style.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
-  <script> scr="js/main.js"</script>
   <script src="js/functions.js"></script>
   <title>Кекстаграм</title>
 </head>
@@ -233,6 +232,8 @@
   <template id="messages">
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
+
+  <script src="js/main.js" type="module"></script>
 
 </body>
 </html>
